+++ go.mod
@@ -4,13 +4,22 @@ go 1.13
 
 require (
 	github.com/fatih/camelcase v1.0.0
+	github.com/fatih/color v1.7.0 // indirect
 	github.com/golang/protobuf v1.3.2
 	github.com/google/go-cmp v0.3.0
+	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
 	github.com/kr/pretty v0.1.0 // indirect
+	github.com/mattn/go-colorable v0.1.2 // indirect
+	github.com/mattn/go-isatty v0.0.9 // indirect
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/sirupsen/logrus v1.4.2 // indirect
+	github.com/spf13/pflag v1.0.3 // indirect
 	go.coder.com/go-tools v0.0.0-20190317003359-0c6a35b74a16
 	go.opencensus.io v0.22.1
 	golang.org/x/lint v0.0.0-20190313153728-d0100b6bd8b3
+	golang.org/x/sync v0.0.0-20190423024810-112230192c58 // indirect
+	golang.org/x/sys v0.0.0-20190904154756-749cb33beabd // indirect
 	golang.org/x/xerrors v0.0.0-20190717185122-a985d3407aa7
-	gotest.tools/gotestsum v0.3.5
+	gotest.tools/gotestsum v0.3.6-0.20190825182939-fc6cb5870c52
 	mvdan.cc/sh v2.6.4+incompatible
 )
