+++ internal/humanfmt/humanfmt.go
@@ -1,4 +1,4 @@
-package console
+package humanfmt
 
 import (
 	"fmt"
