+++ internal/humanfmt/marshal_test.go
@@ -1,4 +1,4 @@
-package console
+package humanfmt
 
 import (
 	"testing"
