+++ test_test.go
@@ -7,13 +7,13 @@ import (
 	"golang.org/x/xerrors"
 
 	"go.coder.com/slog"
-	"go.coder.com/slog/testlog"
+	"go.coder.com/slog/testslog"
 )
 
 func TestExampleTest(t *testing.T) {
 	t.Parallel()
 
-	testlog.Info(t, "my message here",
+	testslog.Info(t, "my message here",
 		slog.F("field_name", "something or the other"),
 		slog.F("some_map", map[string]interface{}{
 			"nested_fields": "wowow",
